refactor html code into separate module
